Organise Views step 1.
